apply code review feedback
