+++ ic-agent/src/agent/agent_error.rs
@@ -1,3 +1,4 @@
+use crate::agent::status::Status;
 use crate::hash_tree::Label;
 use crate::RequestIdError;
 use leb128::read;
@@ -90,8 +91,11 @@ pub enum AgentError {
     #[error("BLS DER-encoded public key is invalid. Expected the following prefix: ${expected:?}, but got ${actual:?}")]
     DerPrefixMismatch { expected: Vec<u8>, actual: Vec<u8> },
 
-    #[error("The status response did not contain a root key")]
-    NoRootKeyInStatus(),
+    #[error("The status response did not contain a root key.  Status: {0}")]
+    NoRootKeyInStatus(Status),
+
+    #[error("Could not read the root key")]
+    CouldNotReadRootKey(),
 
     #[error("Failed to initialize the BLS library")]
     BlsInitializationFailure(),
